Resolve "Bare akseptere tall på input kg/rep"

Closes #104

See merge request gruppe-1/prog2052-prosjekt!70

+++ frontend/app/src/main/java/com/example/workoutapp/data/database/entities/history/SetEntity.kt
@@ -35,6 +35,6 @@ data class SetEntity(
 
     val exerciseEntityId: String,
     val rep: Int,
-    val kg: Int,
+    val kg: Double,
     val typeSet: Int
 )
